requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.10.0
APScheduler>=3.10.0
